phi4-object-detection/requirements.txt
azure-ai-inference>=1.0.0b9
python-dotenv>=1.0.0
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
